requirements.txt
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycodestyle==2.6.0
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
rq==1.7.0
six==1.15.0
toml==0.10.2
twilio==6.52.0
urllib3==1.26.3
Werkzeug==1.0.1
